diffinsights_web: Set ylabel to say the type of contribution
Now you can see wnat type of contribution is displayed on the plot,
instead of having generic "Contributions" label.

This change touched more files than just plots/timerseries.py because
some code reorganization was needed to void circular dependencies.
Therefore some variables and types were moved to views/__init__.py.

The app/contributors.py change is "while at it" code style cleanup.
